Add new page to add pdf preview
