+++ app/assets/scripts/components/documents/single-view/index.js
@@ -1,18 +1,8 @@
 import React, { useCallback, useEffect } from 'react';
-import T from 'prop-types';
 import styled from 'styled-components';
 import { useHistory, useParams } from 'react-router';
 import { GlobalLoading } from '@devseed-ui/global-loading';
-import {
-  glsp,
-  media,
-  multiply,
-  themeVal,
-  listReset
-} from '@devseed-ui/theme-provider';
-import { Heading } from '@devseed-ui/typography';
 import { VerticalDivider } from '@devseed-ui/toolbar';
-import { Button } from '@devseed-ui/button';
 
 import App from '../../common/app';
 import {
@@ -23,18 +13,13 @@ import {
 } from '../../../styles/inpage';
 import UhOh from '../../uhoh';
 import Forbidden from '../../../a11n/forbidden';
-import Prose, { proseSpacing } from '../../../styles/typography/prose';
-import DetailsList from '../../../styles/typography/details-list';
 import DocumentHeadline from '../document-headline';
 import DocumentActionsMenu from '../document-actions-menu';
 import DocumentOutline from './document-outline';
 import DocumentDownloadMenu from '../document-download-menu';
-import DocumentBody from './document-body';
 import { ScrollAnchorProvider } from './scroll-manager';
-import Datetime from '../../common/date';
-import Tip from '../../common/tooltip';
-import { CopyField } from '../../common/copy-field';
 import DocumentGovernanceAction from '../document-governance-action';
+import DocumentContent from './document-content';
 
 import {
   useSingleAtbd,
@@ -49,7 +34,6 @@ import {
 } from '../../../context/comment-center';
 import { useThreadStats } from '../../../context/threads-list';
 import { useEffectPrevious } from '../../../utils/use-effect-previous';
-import { getCitationPublicationDate } from '../citation';
 
 const DocumentCanvas = styled(InpageBody)`
   padding: 0;
@@ -66,83 +50,6 @@ const InpageViewActions = styled(InpageActions)`
   ${VerticalDivider}:first-child {
     display: none;
   }
-`;
-
-const DocumentProse = styled(Prose)`
-  > * {
-    position: relative;
-    padding-bottom: ${multiply(proseSpacing, 2)};
-    margin-bottom: ${multiply(proseSpacing, 2)};
-
-    &::after {
-      position: absolute;
-      bottom: 0;
-      left: 0;
-      height: 1px;
-      width: 100%;
-      background: ${themeVal('color.baseAlphaC')};
-      content: '';
-      pointer-events: none;
-    }
-  }
-
-  > *:last-child {
-    padding-bottom: 0;
-    margin-bottom: 0;
-
-    &::after {
-      display: none;
-    }
-  }
-`;
-
-const DocumentContent = styled.div`
-  max-width: 52rem;
-  padding: ${glsp(themeVal('layout.gap.xsmall'))};
-
-  ${media.smallUp`
-    padding: ${glsp(themeVal('layout.gap.small'))};
-  `}
-
-  ${media.mediumUp`
-    padding: ${glsp(themeVal('layout.gap.medium'))};
-  `}
-
-  ${media.largeUp`
-    padding: ${glsp(themeVal('layout.gap.large'))};
-  `}
-
-  ${media.xlargeUp`
-    padding: ${glsp(themeVal('layout.gap.xlarge'))};
-  `}
-`;
-
-const DocumentHeader = styled.header`
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-gap: ${glsp(2)};
-`;
-
-const DocumentTitle = styled(Heading)`
-  margin: 0;
-`;
-
-const DocumentMetaDetails = styled(DetailsList)`
-  background: transparent;
-`;
-
-const DOIValue = styled.dd`
-  display: flex;
-  align-items: center;
-
-  ${Button} {
-    margin-left: ${glsp(0.25)};
-  }
-`;
-
-const KeywordUnder = styled.small`
-  display: block;
-  opacity: 0.48;
 `;
 
 function DocumentView() {
@@ -291,29 +198,7 @@ function DocumentView() {
           <ScrollAnchorProvider>
             <DocumentCanvas>
               <DocumentOutline atbd={atbd.data} />
-              <DocumentContent>
-                <DocumentProse>
-                  <DocumentHeader>
-                    <DocumentTitle id='doc-header' data-scroll='target'>
-                      {atbd.data.title}
-                    </DocumentTitle>
-                    <DocumentMetaDetails>
-                      <dt>Version</dt>
-                      <dd>{atbd.data.version}</dd>
-                      <ReleaseDate
-                        date={getCitationPublicationDate(atbd.data).date}
-                      />
-                      <DocumentKeywords keywords={atbd.data.keywords} />
-                      <dt>Creators</dt>
-                      <dd>{atbd.data.citation?.creators || 'None provided'}</dd>
-                      <dt>Editors</dt>
-                      <dd>{atbd.data.citation?.editors || 'None provided'}</dd>
-                      <DOIAddress value={atbd.data.doi} />
-                    </DocumentMetaDetails>
-                  </DocumentHeader>
-                  <DocumentBody atbd={atbd.data} />
-                </DocumentProse>
-              </DocumentContent>
+              <DocumentContent atbdData={atbd.data} />
             </DocumentCanvas>
           </ScrollAnchorProvider>
         </Inpage>
@@ -323,121 +208,3 @@ function DocumentView() {
 }
 
 export default DocumentView;
-
-const ReleaseDate = ({ date }) => {
-  if (!date) {
-    return (
-      <React.Fragment>
-        <dt>Release date</dt>
-        <dd>None provided</dd>
-      </React.Fragment>
-    );
-  }
-
-  return (
-    <React.Fragment>
-      <dt>Release date</dt>
-      <dd>
-        <Datetime date={date} />
-      </dd>
-    </React.Fragment>
-  );
-};
-
-ReleaseDate.propTypes = {
-  date: T.object
-};
-
-const DOIAddress = ({ value }) => {
-  if (!value) {
-    return (
-      <React.Fragment>
-        <dt>DOI</dt>
-        <dd>None provided</dd>
-      </React.Fragment>
-    );
-  }
-
-  const isUrl = value.match(/https?:\/\//);
-
-  const doiAddress = isUrl ? value : `https://doi.org/${value}`;
-
-  return (
-    <React.Fragment>
-      <dt>DOI</dt>
-      <DOIValue>
-        <a href={doiAddress} title='DOI of the present document'>
-          {value}
-        </a>
-        <CopyField value={doiAddress}>
-          {({ showCopiedMsg, ref }) => (
-            <Tip
-              title={showCopiedMsg ? 'Copied' : 'Click to copy DOI'}
-              hideOnClick={false}
-            >
-              <Button
-                ref={ref}
-                useIcon='clipboard'
-                variation='base-plain'
-                size='small'
-                hideText
-              >
-                Copy DOI
-              </Button>
-            </Tip>
-          )}
-        </CopyField>
-      </DOIValue>
-    </React.Fragment>
-  );
-};
-
-DOIAddress.propTypes = {
-  value: T.string
-};
-
-const KeywordsList = styled.ul`
-  && {
-    ${listReset()}
-  }
-`;
-
-const DocumentKeywords = ({ keywords }) => {
-  if (!keywords.length) {
-    return (
-      <React.Fragment>
-        <dt>Keywords</dt>
-        <dd>No keywords</dd>
-      </React.Fragment>
-    );
-  }
-
-  return (
-    <React.Fragment>
-      <dt>Keywords</dt>
-      <dd>
-        <KeywordsList>
-          {keywords.map((k) => {
-            const under = k.path.split('|').slice(0, -1).join(' > ');
-            return (
-              <li key={k.value}>
-                {k.label} <KeywordUnder>Under: {under}</KeywordUnder>
-              </li>
-            );
-          })}
-        </KeywordsList>
-      </dd>
-    </React.Fragment>
-  );
-};
-
-DocumentKeywords.propTypes = {
-  keywords: T.arrayOf(
-    T.shape({
-      label: T.string,
-      value: T.string,
-      id: T.int,
-      path: T.string
-    })
-  )
-};
